engine: Move it to the Unity project.
