+++ VisualPinball.Unity/VisualPinball.Unity/Game/DefaultGameEngineAuthoring.cs
@@ -1,5 +1,4 @@
 ﻿using UnityEngine;
-using VisualPinball.Engine.Game.Engine;
 
 namespace VisualPinball.Unity
 {
