+++ VisualPinball.Unity/VisualPinball.Unity/VPT/IApi.cs
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-using VisualPinball.Engine.Game.Engine;
-
 namespace VisualPinball.Unity
 {
 	internal interface IApiInitializable
